Refactor Database God Object into cohesive sub-modules
