+++ .jules/atlas.md
@@ -1,3 +1,7 @@
 ## 2024-03-XX - Module Encapsulation Cleanup
 **Tangle:** Broad visibility (`pub mod`) across many internal modules (`algebra`, `values`, `types`, etc.) leaked implementation details and complicated the dependency graph.
 **Blueprint:** Converted most top-level internal module definitions in `relvar-core`, `relvar-storage`, and `relvar` to `pub(crate) mod`. Re-exported necessary types via their respective `mod.rs` files, ensuring clean, intention-revealing public APIs while maintaining low coupling between internal components.
+
+## 2024-03-XX - Database God Object Breakdown
+**Tangle:** The `Database` struct in `relvar-core/src/database/mod.rs` was a classic God Object anti-pattern (~1000 lines), tangling DDL, DML, constraints, and transactions into a single file. This caused high module coupling and made the codebase difficult to navigate.
+**Blueprint:** Refactored the `Database` impl block into four cohesive sub-modules (`schema.rs`, `data.rs`, `integrity.rs`, `transaction.rs`), each responsible for a distinct domain of database operations. The main `mod.rs` file now acts as a clean Facade, re-exporting only the public API and struct definition.
